pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
fastapi = "^0.112.0"
uvicorn = "0.23.2"
anthropic = "^0.45.2"
requests = "^2.32.3"
openai = "^1.61.1"
python-multipart = "^0.0.20"
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
jinja2 = "^3.1.5"
jose = "^1.0.0"
pycryptodome = "^3.20.0"
starlette = "^0.46.0"
pydantic = "^2.10.6"
bcrypt = "^4.3.0"
pymupdf = "^1.25.4"
markdown = "^3.7"
fitz = "^0.0.1.dev2"
replit-object-storage = "^1.0.2"

[tool.poetry.group.dev.dependencies]

[tool.poetry.extras]

[tool.poetry.group.main.dependencies.pycryptodome]
version = "^3.20.0"
optional = false

# No need to specify pycrypto

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"